feat: graceful shutdown of node envs
