+++ packages/runtime-node/src/index.ts
@@ -21,3 +21,4 @@ export * from './worker-thread-initializer.js';
 export * from './worker-thread-initializer2.js';
 export * from './ws-environment.js';
 export * from './metrics-utils.js';
+export * from './micro-rpc.js';
